feat: 机器人增删改查

+++ model/robot.go
@@ -15,7 +15,7 @@ const (
 
 // Robot 表示微信机器人实例
 type Robot struct {
-	ID           uint           `gorm:"primarykey" json:"id"`
+	ID           int64          `gorm:"primarykey" json:"id"`
 	RobotCode    string         `gorm:"column:robot_code;index;unique,length:64" json:"robot_code"` // 当前机器人的唯一标识
 	Owner        string         `gorm:"column:owner;index;length:64" json:"owner"`                  // 当前机器人的拥有者
 	DeviceId     string         `gorm:"column:device_id;" json:"device_id"`                         // 当前机器人登陆的设备Id
